[FE/style] move service directory
